Fix: FK column/referenced-column type mismatches (INT vs BIGINT UNSIGNED)
Discovered running migrations against ecollab_test after fixing the
whiteboards trailing-comma bug: MySQL/MariaDB errno 150 ("foreign key
constraint incorrectly formed") on the whiteboards table, caused by
whiteboards.channel_id being declared BIGINT UNSIGNED while the column
it references, channels.id, is INT UNSIGNED. FK columns must exactly
match the type of the column they reference.

Ran a full sweep (verified manually against 8 confirmed parent-table id
types: users, channels, servers, messages, subject_classes,
institutions, academic_programs, study_rooms) rather than fixing only
the one table the error message named, since this looked like a
systemic copy-paste pattern rather than an isolated typo. Found and
fixed 5 real instances total, all channel_id/voice_channel_id columns
incorrectly declared BIGINT UNSIGNED against channels.id (INT
UNSIGNED):

  002_core_schema.sql: message_reads.channel_id, whiteboards.channel_id
  006_chat_addon.sql:  message_reads.channel_id, whiteboards.channel_id
                       (redundant duplicate definitions of the same two
                       tables, currently masked by this file's own
                       `SET foreign_key_checks = 0/1` wrapper -- fixed
                       for consistency even though not the proximate
                       cause of any error)
  007_voice_presence.sql: users.voice_channel_id (would have been the
                       next failure after whiteboards, since this file
                       does NOT disable foreign_key_checks)

Also investigated 5 flagged `room_id` columns referencing study_rooms.id
-- these turned out to be a false positive from an early, buggy version
of the scan script (which conflated table names across files). Verified
manually: room_id (INT UNSIGNED) correctly matches study_rooms.id (INT
UNSIGNED) at its first, actually-used definition. No fix needed there.

Noted but NOT fixed (non-blocking, separate issue): study_rooms is
defined twice in 002_core_schema.sql with conflicting id types (INT
UNSIGNED at its first definition, BIGINT UNSIGNED at a second, later
`IF NOT EXISTS` redefinition that silently no-ops since the table
already exists by then). Harmless today only because of IF NOT EXISTS;
worth cleaning up separately as its own task.

+++ database/migrations/006_chat_addon.sql
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS `message_attachments` (
 -- ── message_reads ──────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS `message_reads` (
   `user_id`      BIGINT UNSIGNED NOT NULL,
-  `channel_id`   BIGINT UNSIGNED NOT NULL,
+  `channel_id`   INT UNSIGNED    NOT NULL,
   `last_read_at` DATETIME        NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
   PRIMARY KEY (`user_id`, `channel_id`),
   KEY `idx_channel_id` (`channel_id`),
@@ -50,7 +50,7 @@ CREATE TABLE IF NOT EXISTS `message_reads` (
 -- ── whiteboards ────────────────────────────────────────────
 CREATE TABLE IF NOT EXISTS `whiteboards` (
   `id`          BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
-  `channel_id`  BIGINT UNSIGNED NOT NULL,
+  `channel_id`  INT UNSIGNED    NOT NULL,
   `state_json`  LONGTEXT        NOT NULL,
   `created_by`  BIGINT UNSIGNED NOT NULL,
   `updated_by`  BIGINT UNSIGNED     NULL DEFAULT NULL,
